refactor: moving the datalake validation helpers into the datalake validation package
