+++ azurerm/internal/services/datalake/data_lake_analytics_account_resource.go
@@ -14,6 +14,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/features"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/datalake/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tags"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
@@ -42,7 +43,7 @@ func resourceArmDataLakeAnalyticsAccount() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateDataLakeAccountName(),
+				ValidateFunc: validate.DataLakeAccountName(),
 			},
 
 			"location": azure.SchemaLocation(),
@@ -71,7 +72,7 @@ func resourceArmDataLakeAnalyticsAccount() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateDataLakeAccountName(),
+				ValidateFunc: validate.DataLakeAccountName(),
 			},
 
 			"tags": tags.Schema(),
