+++ azurerm/internal/services/datalake/data_lake_store_firewall_rule_resource.go
@@ -10,8 +10,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/validate"
+	commonValidate "github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/datalake/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
@@ -38,14 +39,14 @@ func resourceArmDataLakeStoreFirewallRule() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateDataLakeFirewallRuleName(),
+				ValidateFunc: validate.DataLakeFirewallRuleName(),
 			},
 
 			"account_name": {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateDataLakeAccountName(),
+				ValidateFunc: validate.DataLakeAccountName(),
 			},
 
 			"resource_group_name": azure.SchemaResourceGroupName(),
@@ -53,13 +54,13 @@ func resourceArmDataLakeStoreFirewallRule() *schema.Resource {
 			"start_ip_address": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validate.IPv4Address,
+				ValidateFunc: commonValidate.IPv4Address,
 			},
 
 			"end_ip_address": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validate.IPv4Address,
+				ValidateFunc: commonValidate.IPv4Address,
 			},
 		},
 	}
